Explorer16 port update

+++ ports/Microchip/Explorer16/PIC24/api/adc/read.c.v
@@ -8,4 +8,4 @@ module adc
 
 // #define adc__read(CHANNEL) ({ AD1CHS = CHANNEL; AD1CON1bits.SAMP = 1; while (!AD1CON1bits.DONE); ADC1BUF0; })
 
-fn C.read(channel int) int 
+fn C.read(channel u8) u16
